+++ src/main/java/me/towdium/jecalculation/gui/Resource.java
@@ -162,6 +162,7 @@ public class Resource {
     public static final ResourceGroup ICN_LIST = new ResourceGroup(ICN_LIST_N, ICN_LIST_F);
     public static final ResourceGroup ICN_LABEL = new ResourceGroup(ICN_LABEL_N, ICN_LABEL_F);
     public static final ResourceGroup ICN_TEXT = new ResourceGroup(ICN_TEXT_N, ICN_TEXT_F);
+    public static final ResourceGroup ICN_HELP = new ResourceGroup(ICN_HELP_N, ICN_HELP_F);
 
     protected int xPos, yPos, xSize, ySize;
     protected ResourceLocation rl;