+++ CHANGELOG.md
@@ -19,3 +19,14 @@ Fixed ItemStack incorrectly converted and rendered
 Fixed Crash caused by containers not synchronized  
 Added Mouse scroll support for help documents  
 Changed Empty OreDict entries will not be displayed
+
+#### 3.0.0-pre15
+
+Added Suggestions in calculator when dest has no recipe  
+Added Tooltip for fuzzy itemStack labels  
+Changed Prohibit adding recipes duplicating existing ones  
+Fixed Amount tooltip for fluid labels  
+Fixed Amount rendering in unicode environment  
+Fixed Right click to pick item from JEI item list    
+Updated Forge version  
+